# Back-End/swagger.yaml
openapi: 3.0.0
info:
  title: EduFlex API
  contact: {}
  version: '1.0'
servers:
- url: https://eduflex-0o5p.onrender.com/api/v1
  variables: {}
paths:
...
  /users:
    get:
      tags:
      - User
      summary: Get All Users
      operationId: GetAllUsers
      parameters:
      - name: role
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: Student
      - name: Cache-Control
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: no-cache
      - name: Postman-Token
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: Host
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: User-Agent
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: PostmanRuntime/7.32.1
      - name: Accept
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '*/*'
      - name: Accept-Encoding
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: gzip, deflate, br
      - name: Connection
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: keep-alive
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /users/show-me:
    get:
      tags:
      - User
      summary: Show Current User
      operationId: ShowCurrentUser
      parameters:
      - name: Cache-Control
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: no-cache
      - name: Postman-Token
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: Host
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: User-Agent
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: PostmanRuntime/7.32.1
      - name: Accept
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '*/*'
      - name: Accept-Encoding
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: gzip, deflate, br
      - name: Connection
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: keep-alive
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - User
      summary: Get Single User
      operationId: GetSingleUser
      parameters:
      - name: Cache-Control
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: no-cache
      - name: Postman-Token
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: Host
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: <calculated when request is sent>
      - name: User-Agent
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: PostmanRuntime/7.32.1
      - name: Accept
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '*/*'
      - name: Accept-Encoding
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: gzip, deflate, br
      - name: Connection
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: keep-alive
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...